+++ testql/interpreter/converter/__init__.py
@@ -1,0 +1,19 @@
+"""IQL/TQL to TestTOON converter.
+
+Converts IQL/TQL scripts to TestTOON format (*.testql.toon.yaml).
+
+Usage:
+    from testql.interpreter.converter import convert_iql_to_testtoon
+    result = convert_iql_to_testtoon(source_text, filename)
+"""
+
+from .core import convert_iql_to_testtoon, convert_file, convert_directory
+from .models import Row, Section
+
+__all__ = [
+    "convert_iql_to_testtoon",
+    "convert_file",
+    "convert_directory",
+    "Row",
+    "Section",
+]